requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
paho-mqtt
ucimlrepo
joblib
streamlit